cleanup image-flashcards and j2 templates

+++ image-flashcards.sh
@@ -1,83 +1,84 @@
 #!/bin/bash
 
+if [[ $@ =~ "-d" ]] || [[ $@ =~ "--debug" ]]; then
+  set -x
+fi
+
 if [[ $@ =~ "-h" ]] || [[ -z "$@" ]]; then
   echo "argument help:
 
-  exsample:   
-  $0 master dev=/dev/sda
-  $0 master dev=/dev/sda --copy-cfg
-  
+  exsample: 
+  $0 image dev=/dev/sda
+  $0 image dev=/dev/sda --copy-cfg
+
   -d #  debug
   -skip-dd #  skip dd image
   -copy-cfg #  copy configs
-  H#1 #  hoatname post fix number
 
   Arg: 1
-  master #  write master iamge to card
-  worker #  write worker iamge to card
-  Arg: 2  
+  image#  write iamge to card
+  
+  Arg: 2
   dev=/dev/sda #  cf card device
 
-  Arg: 3  
-  image=os.img #  image file 
+  Arg: 3
+  image=os.img #  image file
+
+  Arg: 4
+  ip=10.5.94.51 # set fixed address
+
+  Arg: 5
+  shortname=master-01 # set hostname 
   "
 fi
 
 cmd_args=$@
 
+if [[ ! -f /usr/bin/j2 ]]; then
+  echo "Missing /usr/bin/j2, Enter or CTRL-c ro run or quit: apt install j2cli"
+  read
+  apt install j2cli
+  exit 1
+fi
+
 if [[ ! -f /usr/bin/pv ]]; then
-  echo "Missing /usr/bin/pv, Enter or CTRL-c ro run or quit: apt install pv" 
+  echo "Missing /usr/bin/pv, Enter or CTRL-c ro run or quit: apt install pv"
   read
   apt install pv
   exit 1
-fi 
-
-if [[ $@ =~ "--skip-dd" ]]; then
-  skip_dd=true	
 fi
 
-if [[ $@ =~ "H#" ]]; then
-  hostpostfix=$(echo $@| awk -F '#' '{print $2}')
-fi 
+if [[ $@ =~ "--skip-dd" ]]; then
+  skip_dd=true
+fi
+
 
 if [[ -z "$2" ]]; then
   echo "missing device"
   exit 1
 else
   DEV=$(echo $2| grep dev| awk -F = '{print $2}')
-fi 
+fi
 
 if [[ -z "$3" ]]; then
-  echo "missing image file" 
+  echo "missing image file"
   exit 1
 else
   IMAGE=$(echo $3| grep image| awk -F = '{print $2}')
 fi
 
+if [[ -z "$4" ]]; then
+  echo "missing ip"
+  exit 1
+else
+  ip=$(echo $4| grep ip| awk -F = '{print $2}')
+fi
 
-post(){
-  boot="/media/${USER}/HypriotOS"
-  root="/media/${USER}/root"
-  set -x
-  touch $boot/ssh
-
-  if [[ $1 == "master" ]]; then
-    echo "master file copy"
-    cp user-data.master $boot/user-data
-    cp config-master.txt $boot/config.txt
-  fi
-  if [[ $1 == "worker" ]]; then
-    echo "worker file copy"
-    cp user-data${hostpostfix}.worker $boot/user-data
-    cp config-worker${hostpostfix}.txt $boot/config.txt
-  fi
-  un_mount
-  set +x
-}
-
-if [[ $@ =~ "--copy-cfg" ]]; then
-  post $1
-  exit 0  
+if [[ -z "$5" ]]; then
+  echo "missing shortname"
+  exit 1
+else
+  shortname=$(echo $5| grep shortname| awk -F = '{print $2}')
 fi
 
 function is_mount() {
@@ -107,56 +108,51 @@ function is_mount() {
 }
 
 un_mount(){
-  set -x	
   if [[ "$mount_status" == "yes" ]]; then
-    	  
+
     $(sudo umount ${mount_path1} )
     $(sudo umount ${mount_path2} )
-  fi   
-  set +x	
+  fi
 }
+
+post(){
+  is_mount
+  boot="/media/${USER}/HypriotOS"
+  root="/media/${USER}/root"
+  touch $boot/ssh
+  user_data_templated=`echo inventory_hostname_short=$shortname ansible_host=$ip | j2 --format=env roles/bootstrap-core/templates/user-data.j2`
+  echo "$user_data_templated" > $boot/user-data
+  cp roles/bootstrap-core/files/config.txt $boot/config.txt
+  cp roles/bootstrap-core/files/cmdline.txt $boot/cmdline.txt
+  un_mount
+  exit 0
+}
+
+if [[ $@ =~ "--copy-cfg" ]]; then
+  post $1
+  exit 0
+fi
 
 echo $@
 is_mount $@
 
-if [[ $@ =~ "master" ]]; then
+if [[ $@ =~ "image" ]]; then
   if [[ ${os_found} == "rasp" ]] || [[ ${write_status} == "force" ]]; then
-    echo "writing master image ${IMAGE} to device ${DEV}"
+    echo "writing image ${IMAGE} to device ${DEV}"
     echo "Are you sure you want to write to the disk at $DEV: CTRL-C to quit"
     read
-    if [[ ${skip_dd} == "true" ]]; then  
-      echo "skipping dd"   	    
-    else	    
-    	$(/usr/bin/pv ${IMAGE} | sudo dd of=${DEV})
+    if [[ ${skip_dd} == "true" ]]; then
+      echo "skipping dd"
+    else
+      $(/usr/bin/pv ${IMAGE} | sudo dd of=${DEV})
     fi
     un_mount
-    echo "ejected fc cards, please reinsert" 
+    echo "ejected fc cards, please reinsert"
     read
     post  $1
     exit 0
   else
     echo "already found $os_found"
     exit 1
-  fi 
-fi 
-
-if [[ $@ =~ "worker" ]]; then
-  if [[ ${os_found} == "rasp" ]] || [[ ${write_status} == "force" ]]; then
-    echo "writing worker image ${IMAGE} to device ${DEV}"
-    echo "Are you sure you want to write to the disk at $DEV: CTRL-C to quit"
-    read
-    if [[ ${skip_dd} == "true" ]]; then  
-       echo "skipping dd"   	    
-    else	    
-       $(/usr/bin/pv ${IMAGE} | sudo dd of=${DEV})
-    fi
-    un_mount
-    echo "ejected fc cards, please reinsert" 
-    read
-    post  $1
-    exit 0
-  else
-    echo "already found $os_foud"
-    exit 1
-  fi 
-fi 
+  fi
+fi
